Add clientside callbacks chapter.
