+++ dash_docs/tutorial/__init__.py
@@ -2,6 +2,7 @@ import os
 from .import advanced_callbacks
 from .import auth
 from .import callback_gotchas
+from .import clientside_callbacks
 from .import core_component_examples as examples
 from .import core_components
 from .import cytoscape
